refactor(api): déplacer getComments vers shantytownCommentModel.findByShantytown
refers to https://trello.com/c/ZPU4RYve/2668
