+++ packages/api/server/models/shantytownCommentModel/index.ts
@@ -1,5 +1,6 @@
 import create from './create';
 import findAll from './findAll';
+import findByShantytown from './findByShantytown';
 import findOne from './findOne';
 import getHistory from './getHistory';
 import deleteComment from './deleteComment';
@@ -7,6 +8,7 @@ import deleteComment from './deleteComment';
 export default {
     create,
     findAll,
+    findByShantytown,
     findOne,
     getHistory,
     deleteComment,
